+++ src/main/java/com/alivro/spring/sleepyringtail/dao/CategoryDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryDao extends JpaRepository<Category, Long> {
+public interface CategoryDao extends JpaRepository<Category, Integer> {
     /**
      * Método para buscar la existencia de una categoría por su nombre
      *
